requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pymongo==4.7.2
python-dotenv==1.0.1
